refactor(env): Extract Prototypes into separate files
