+++ src/api/kuma.js
@@ -1,0 +1,13 @@
+/**
+ * @prettier
+ */
+const url = require('url');
+const util = require('./util.js');
+
+module.exports = {
+    /**
+     * Expose url from node.js to templates
+     */
+    url: url,
+    htmlEscape: util.htmlEscape
+};
