+++ src/api/uri.js
@@ -1,0 +1,9 @@
+/**
+ * @prettier
+ */
+module.exports = {
+    // Encode text as a URI component.
+    encode(str) {
+        return encodeURI(str);
+    }
+};
